[ROS2][fix] Fix DeviceId

+++ lidar_conversions/src/LivoxToLidarNode.h
@@ -77,7 +77,7 @@ private:
   rclcpp::Subscription<LivoxCustomMsg>::SharedPtr LivoxMsgListener;
   rclcpp::Publisher<Pcl2_msg>::SharedPtr Talker;
 
-  std::map<std::string, int> DeviceIdMap;  ///< Map to store the device id of each device (in case of multilidar).
+  std::map<std::string, uint8_t> DeviceIdMap;  ///< Map to store the device id of each device (in case of multilidar).
 
   bool IsPcl2 = false;
 };
